+++ CHANGELOG.md
@@ -5,8 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
 ### Addded
+
 ### Changed
+
+* (`fitsio`) _breaking change_: inverted the order of image/region axes to match the C row-major convention
+
 ### Removed
 
 ## [0.12.1] - 2018-02-23